Add basic Time Matrix layout
Added a basic time matrix layout that consists of four fragments with a
tab bar for the user to either click or swipe between. This is a
rudimentary layout to show how the time matrix should be presented. In
order to accomplish this, the Android Design Support Library was added
as a dependency.