Connect AddItemDialogFragment to QuadrantFragment
Added a callback listener for the AddItemDialogFragment to pass the
necessary information back to instantiate a new QuadrantFragment and add
it to the current QuadrantFragment.

Listener was appended to the AddItemOnClickListener for sake of
simplicity; however, it should be refactored in the future to adhere
more to the single responsibility principle.