Implement rudimentary read for SQL DB
This commits include very rough code. It is only meant as a proof of
concept. This code will require refactoring before it is properly
incorporated into the development branch.

While working on this branch, a flaw in the save implementation was
discovered. The save implementation needs to be rewritten to include
updating existing rows.