Read Matrix data
Matrix data can now be loaded on start up. Any item added to a matrix
will now be saved to the SQL DB, and loaded on app startup.

Additionally, in the AddItemOnClickListener, QuadrantFragment has been
substituted for its abstract form in order to follow more closely to the
SOLID design philosophy.