Fix remove item reappearance bug
Fixed the bug where deleted items would reappear if the app was
minimized then resumed. The bug appeared because the data storage wasn't
being informed of the deletion. Deletion was occuring in the view space
and changes were not being propagated to the model space.

Deletion was moved to higher in the view space where the presenter space
could update the model space. This resolved the bug.

Additionally, the EditItemDialogFragmentTest needed to be modified to
handle the changes. The resulting tests are more simplified than the
previous tests; however, there are areas that could be improved.