Fix Issue #18
Fixed Issue #18 (https://github.com/Austin-Ray/Hexis/issues/18). The
issue did not result from the add item code as previously suspected, but
rather from the code that inflated the ViewHolder.

The issue was due to the ViewHolder being inflated to the same size as
the RecyclerView. The additional items were there; however, since each
ViewHolder was as large as the RecyclerView the other items did not
appear to exist. If you scrolled the RecyclerView you could find the
other items.