Refactor QuadrantItem SQL code
Refactored QuadrantItemReader and QuadrantItemWriter to inherit from
QuadrantItemSqlInterator. The reader and writer now share some methods
that are used by both. Previous implementation had the writer
instantiating a reader to check if it exists.

The main goal of the refactor was to reduce the amount of code to
prevent use of case dependent methods. Lower level code was implemented
to update an SQL row given an item, quadrant number, and/or goal number.
Furthermore, querying information in a column can be down via higher
level methods that call a lower level method. The lower level method
will return whatever type in the that position; however, casting is
necessary.

SQL code for GoalReader and GoalWriter needs to be refactored, but was
not touched in this refactoring period due to not being used.

Static analysis was used on all the SQL code.