Implement changing existing items' quadrant
Created a rough implementation for moving QuadrantItems between
quadrants. In order to accomplish this, MVP had to be broken slightly to
allow a IQuadrantPresenter to fetch data from IMatrixPresenter.
Otherwise, the Quadrant layer of code would not be able to swap between
fragments.

Code added in this commit needs to be refactored to reduce its length
and to simplify logic. This is not final code and will require a
refactor being being merged into develop