Increase coverage for io.ray.hexis.view to 96%
Increased the unit test coverage for the io.ray.hexis.view package to
96% according to the local JaCoCo report.

Discrepancies between CodeCov and local reports are bound to occur This
commit includes a new Gradle task to reduce the variation in coverage
percentages. To use the new report run jacocoDebugReport or
jacocoReleaseReport and open the HTML file in app/build/reports/jacoco
to view the results. I am currently exploring a way have to report
appear in Android Studio.